wip: Improving tests
